+++ src/Pages/JoinPage/JoinPage.tsx
@@ -1,5 +1,5 @@
 import HeaderLogo from "../../components/common/JoinLogo/JoinLogo";
-import { JoinForm } from "../../components/Join/JoinForm/JoinForm";
+import { JoinForm } from "../../components/auth/JoinForm/JoinForm";
 
 function JoinPage() {
   return (
