+++ src/components/auth/JoinInput/JoinInput.tsx
