+++ src/components/auth/LoginForm/style.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import Button from "../common/Button/Button";
+import Button from "../../common/Button/Button";
 
 export const LoginForm = styled.form`
   width: 55rem;
